fix: frontend test
